sample rate = ADC per 4ms , sample AN0 per 40ms AN2SampValue[AN2_ARY_SAMPLE_SIZE] only sample per 5*4ms CheckValResonable() / OVER_TH_LIMIT=100 AN2 avg 5 time(AN2_AVG_5_CHK_BUF_FUN)

+++ HM_HeartRate.c
@@ -26,7 +26,7 @@ uint8_t AGC_MCP4011_GainLimitSet,AN2_GainContOutOfRange_CycleCnt;
 
 uint8_t InRangeStatus,AN2_HRBufferIndex;
 
-uint8_t AN2_oldPulseCnt;
+uint16_t AN2_oldPulseCnt;
 
 uint8_t		PIC_HR_Msg_OUT[SEND_toRTL_HR_MSG_SIZE] = {0xff,'H', '=' ,'1','N','S','1','\n' , '\r'} ;
